2.5.21

+++ Graph/TimeSeries/graphTimeSeriesTimeFrequency.m
@@ -70,7 +70,7 @@ yPcolor = freq.data;
 % xPcolor = [time.data(1:end-1) - diff(time.data)/2; time.data(end) - (time.data(end)-time.data(end-1))/2];
 % yPcolor = [freq.data(1:end-1) - diff(freq.data)/2; freq.data(end) - (freq.data(end)-freq.data(end-1))/2];
 
-h = pcolor(ax, xPcolor, yPcolor, double(var.data'));
+h = pcolor(ax, double(xPcolor), double(yPcolor), double(var.data'));
 set(h, 'FaceColor', 'flat', 'EdgeColor', 'none');
 cb = colorbar();
